refactoring(integrations): better folder structure
