+++ src/components/IntegrationsView/Header/IntegrationsHeader.tsx
@@ -1,12 +1,12 @@
 import { Button, Flex, Heading, IconButton } from "@radix-ui/themes";
-import { useWindowDimensions } from "../../hooks/useWindowDimensions";
+import { useWindowDimensions } from "../../../hooks/useWindowDimensions.ts";
 import type { FC } from "react";
 import { ArrowLeftIcon } from "@radix-ui/react-icons";
 import styles from "./IntegrationsHeader.module.css";
-import { LeftRightPadding } from "../../features/Integrations/Integrations";
-import { toPascalCase } from "../../utils/toPascalCase";
-import { useAppSelector } from "../../hooks";
-import { selectConfig } from "../../features/Config/configSlice.ts";
+import { LeftRightPadding } from "../../../features/Integrations/Integrations.tsx";
+import { toPascalCase } from "../../../utils/toPascalCase.ts";
+import { useAppSelector } from "../../../hooks/index.ts";
+import { selectConfig } from "../../../features/Config/configSlice.ts";
 
 type IntegrationsHeaderProps = {
   handleFormReturn: () => void;
